requirements.txt
fastapi[all] 
sqlalchemy 
psycopg2 
alembic 
python-dotenv
firebase_admin